requirements.txt
numpy
pytorch
qpth
block